requirements.txt
langchain==0.1.20
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.34.0
faiss-cpu==1.8.0
streamlit-extras==0.4.2